backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-socketio==5.1.1
opencv-python==4.5.3.56
numpy==1.21.2
pyyaml==5.4.1
python-engineio==4.2.1
python-socketio==5.4.0
